Fix: 	Hostname string length to avoid buffer overrun. 	Print PSK and SSID over bluetooth when submitted for configuration.
Don't print the OTA upload progress over serial to improve upload speed.

Add a timestamp to the ntfy message so that the local time of the feeder
notification is in the message. The ntfy app uses the phone's local time
for the notification time which may be in a different timezone to the
feeder.

Remove the large malloc() in favour of client.printf() to reduce the
runtime memory footprint.